requirements.txt
# Core dependencies (minimal set)
jinja2>=3.1.2
pyyaml>=6.0
jsonschema>=4.17.3

# Date/time handling
python-dateutil>=2.8.2
pytz>=2023.3

# Document generation
markdown>=3.4.3